+++ src/amulet/app/exception/__init__.py
@@ -1,4 +1,4 @@
-from ._api import (
+from ._exception import (
     display_exception,
     display_exception_blocking,
     CatchExceptionDialog,
