+++ src/amulet/app/exception/_exception.py
